# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TicketflowAI API",
        "version": "0.5"
    },
    "paths": {
...
        "/intervals": {
            "get": {
                "tags": [
                    "Intervals"
                ],
                "summary": "Get a list of intervals",
                "operationId": "b3c2b6306ca2993bf2d7eeca6fea3123",
                "responses": {
                    "200": {
                        "description": "List of intervals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/IntervalResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Intervals"
                ],
                "summary": "Create a new interval",
                "operationId": "49922ac2e96a30e322b125a7a922d91b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "details"
                                ],
                                "properties": {
                                    "name": {
                                        "example": "Weekly"
                                    },
                                    "details": {
                                        "example": "Every 7 days"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Interval created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IntervalResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/intervals/{id}": {
            "get": {
                "tags": [
                    "Intervals"
                ],
                "summary": "Get details of a specific interval",
                "operationId": "eff07503cbc9d7cf14727c6e3b137517",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the interval",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Interval details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IntervalResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Interval not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Intervals"
                ],
                "summary": "Update an interval",
                "operationId": "3ff8e50c2402a00bf79121acbb0c39ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the interval",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "details"
                                ],
                                "properties": {
                                    "name": {
                                        "example": "Updated Interval"
                                    },
                                    "details": {
                                        "example": "Updated details"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Interval updated successfully"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Intervals"
                ],
                "summary": "Delete an interval",
                "operationId": "98c9593305e7a2bd87e6c9f14eb1256b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the interval",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Interval deleted successfully"
                    },
                    "404": {
                        "description": "Interval not found"
                    }
                }
            }
        },
        "/messages": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Create a new message for a ticket",
                "operationId": "941d583a09ed64becb2c2f4753c22501",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ticket_id",
                                    "content"
                                ],
                                "properties": {
                                    "ticket_id": {
                                        "example": 123
                                    },
                                    "content": {
                                        "example": "This is a message content."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/messages/{id}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Get messages for a specific ticket",
                "operationId": "abaddc21ac1f3e46b3cf1527df9bad66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the message",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MessageResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Message not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "IntervalResource": {
                "properties": {
                    "id": {
                        "description": "Unique identifier for the interval",
                        "type": "integer",
                        "example": 1
                    },
                    "days": {
                        "description": "Number of days for the interval",
                        "type": "integer",
                        "example": 7
                    },
                    "type": {
                        "description": "Type of interval (e.g., weekly, monthly)",
                        "type": "string",
                        "example": "Weekly"
                    },
                    "template_name": {
                        "description": "Name of the associated template",
                        "type": "string",
                        "example": "Weekly Report Template"
                    },
                    "subject_template": {
                        "description": "Subject template for the interval",
                        "type": "string",
                        "example": "Weekly Summary"
                    }
                },
                "type": "object"
            },
            "MessageResource": {
                "properties": {
                    "id": {
                        "description": "Unique identifier for the message",
                        "type": "integer",
                        "example": 1
                    },
                    "ticket_id": {
                        "description": "ID of the associated ticket",
                        "type": "integer",
                        "example": 123
                    },
                    "content": {
                        "description": "Content of the message",
                        "type": "string",
                        "example": "This is a message content."
                    },
                    "user_id": {
                        "description": "ID of the user who created the message",
                        "type": "integer",
                        "example": 45
                    },
                    "user_name": {
                        "description": "First name of the user who created the message",
                        "type": "string",
                        "example": "John"
                    },
                    "user_lastname": {
                        "description": "Last name of the user who created the message",
                        "type": "string",
                        "example": "Doe"
                    },
                    "user_role": {
                        "description": "Role of the user who created the message",
                        "type": "string",
                        "example": "technician"
                    },
                    "timestamp": {
                        "description": "Timestamp of when the message was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}